+++ config/runtime-electron-baseline.txt
@@ -5,9 +5,6 @@
 
 src/main/browser/cdp-bridge.ts
 src/main/ipc/filesystem-watcher.ts
-src/main/ipc/plugin-marketplaces.ts
-src/main/ipc/plugins.ts
 src/main/jira/authenticated-request.ts
 src/main/network/proxy-settings.ts
 src/main/speech/model-manager.ts
-src/main/speech/stt-service.ts
